Started refactoring compression in repair
Problem: Refactoring the old implementation that used lists

Solution: Using classes from graph and node and python's priority class.
Currently trying to find a way to find the current freq in update
dictionary so that it can be updated or replaced. Priority queue doesn't
have an iterator.

Might extend the base class and override the put method

[ticket: 10]