Introducing clusters  (#37)
* Introducing Clusters again

Creating a list of graphs was not enough because repair wouldn't see all the nodes
This would have ended up being a huge problem because it will either compress
one graph at a time (no point in creating clusters in this case) or compressing nothing
at all/getting an error of sorts).

Introduced clusters again. This solves the problem  because it partially impelments the
idea of a single graph (e.g. provides a `list_nodes` for adj. representation, provides add_node method, and also inherits from Graph). But it also doesn't change the nodes in anyway, for example the `graph_id` of nodes stays the same. This means the graph-node owenership is not affected despite having one central cluster that acts like a graph. Communication with the grpah is then handled through events. See the related issue for more info.

Issue 33

* Merged duplicate files. Tests should be under tests

* last duplicate clean up is failing. Avoiding it for now.

* Sort the list the first time to be sure.

* fixed failing test bc it was assuming position of nodes.