Working Implementation of generator  (#43)
* rewriting the generator. Tests are infinte looping.

* fixed the infinite loop issue the tests picked up on. The while loop was marking graphs with true without incrementing the counter twice. Fixed by adjusting the iterator control logic.

* Refactored the solution to clean up some of the logic and make it more clear. Extracted the statements inside the while loop into a function and split the conditions into two conditions.

* Basic implementation of the graph generator

This issue has been around for far too long and I finally got around to it today. This provides a basic version of the generator with minimal testing. It doesn't inf loop and it doesn't break randomly so it's a good step forward.

The implementation is very similar to the last two iterations the main difference is that I have cleaned up the code and split it into many functions. This should make debugging it easier whenever it breaks.

We can close Issue #18 now since we have the basics implemented. Track bugs in d/f issues as they come up.